Move interfaces to types folder
As suggested.
